zip 3.1d25_l_s6x Beta
Origin: http://antinode.info/ftp/info-zip//zip31d25_l_s6x.zip

+++ win32/win32.c
@@ -1302,7 +1302,7 @@ int gettimeofday(struct timevalstruct *tv, struct timezonestruct *tz)
   {
     GetSystemTimeAsFileTime(&ft);
 
-    tmpres = ft.dwHighDateTime << 32;
+    tmpres = (__int64)ft.dwHighDateTime << 32;
     tmpres |= ft.dwLowDateTime;
 
     tmpres /= 10;  /*convert into microseconds*/
